these tests actually don't have to be generated
put them in a separate file
